refactor: rename module to identityServiceHelper
